cambio de float a double

+++ src/interfaz/controladorInterfaz.java
@@ -113,9 +113,9 @@ public class controladorInterfaz implements Initializable{
 		
 		mm.textProperty().addListener((observable,lastValue,newValue)->{
 			try{
-				float ini=Float.parseFloat(((TextField)tabla.getChildren().get(7)).getText());
+				double ini=Double.parseDouble(((TextField)tabla.getChildren().get(7)).getText());
 				System.out.println(ini);
-				float rest=ini-(Float.parseFloat(mm.getText()));
+				double rest=ini-(Double.parseDouble(mm.getText()));
 				((TextField)tabla.getChildren().get(tam-2)).setText(String.valueOf(rest));
 				((TextField)tabla.getChildren().get(tam-2)).setEditable(false);
 				((TextField)tabla.getChildren().get(tam-1)).setEditable(false);
@@ -144,8 +144,8 @@ public class controladorInterfaz implements Initializable{
 				//int cantidad=((tam-16)/7)+1;
 				//System.out.println(cantidad);
 				for (int i=14; i<tabla.getChildren().size();i+=6){
-					float m=Float.parseFloat(((TextField)tabla.getChildren().get(i)).getText());
-					float df=Float.parseFloat(((TextField)tabla.getChildren().get(i+3)).getText());
+					double m=Double.parseDouble(((TextField)tabla.getChildren().get(i)).getText());
+					double df=Double.parseDouble(((TextField)tabla.getChildren().get(i+3)).getText());
 					switch(carga.getId()){
 						case "carga1":	s1.getData().add(new XYChart.Data(m,df));break;
 						case "carga2":	s2.getData().add(new XYChart.Data(m,df));break;
@@ -181,8 +181,8 @@ public class controladorInterfaz implements Initializable{
 				//int cantidad=((tam-16)/7)+1;
 				//System.out.println(cantidad);
 				for (int i=14; i<tabla.getChildren().size();i+=6){
-					float m=Float.parseFloat(((TextField)tabla.getChildren().get(i)).getText());
-					float df=Float.parseFloat(((TextField)tabla.getChildren().get(i+3)).getText());
+					double m=Double.parseDouble(((TextField)tabla.getChildren().get(i)).getText());
+					double df=Double.parseDouble(((TextField)tabla.getChildren().get(i+3)).getText());
 					switch(carga.getId()){
 						case "carga1":	s1.getData().add(new XYChart.Data(m,df));break;
 						case "carga2":	s2.getData().add(new XYChart.Data(m,df));break;
@@ -465,11 +465,11 @@ public class controladorInterfaz implements Initializable{
 				
 	}
 	
-	float []consta= new float[5]; //0>diam 1>alt 2>ar 3>vol 4>pesoAnillo
-	float []humIni= new float[4]; //0>wm 1>ws 2>wf 3>wpercent
-	float [] antPrueba=new float[10];//0>wma 1>wmnatural 2>humedad 3>wmnaturals 4>volmnatural 5>wamnatural 6>voltmnatural 7>volvnatural 8>vacios 9>gradSat
-	float [] despPrueba= new float[11];//0>pesomanillo 1>Wm 2>pesomseca 3>Ws 4>W% 5>Vm 6>Vs 7>Vv 8>e 9>G% 10>Ww 
-	final float PI=3.14159265f;
+	double []consta= new double[5]; //0>diam 1>alt 2>ar 3>vol 4>pesoAnillo
+	double []humIni= new double[4]; //0>wm 1>ws 2>wf 3>wpercent
+	double [] antPrueba=new double[10];//0>wma 1>wmnatural 2>humedad 3>wmnaturals 4>volmnatural 5>wamnatural 6>voltmnatural 7>volvnatural 8>vacios 9>gradSat
+	double [] despPrueba= new double[11];//0>pesomanillo 1>Wm 2>pesomseca 3>Ws 4>W% 5>Vm 6>Vs 7>Vv 8>e 9>G% 10>Ww 
+	final double PI=3.14159265f;
 	
 @FXML
 public void initialize() {
@@ -522,7 +522,7 @@ public void initialize() {
 		
 		Ss.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float ss=Float.parseFloat(((String)Ss.getText()));
+				double ss=Double.parseDouble(((String)Ss.getText()));
 				antPrueba[4]=antPrueba[3]/ss;
 				volmnatural.setText(String.valueOf(antPrueba[4]));
 				
@@ -539,7 +539,7 @@ public void initialize() {
 		//para calcular las constantes de equipo/////////////////////////
 		diametro.textProperty().addListener((observable, oldValue, newValue) -> {
 		    try{
-		    	consta[0]=Float.parseFloat(((String)diametro.getText()));	
+		    	consta[0]=Double.parseDouble(((String)diametro.getText()));	
 		    	consta[2]=PI*consta[0]*consta[0]/4; 
 		    	consta[3]=consta[1]*consta[2];
 				area.setText(String.valueOf(consta[2]));
@@ -555,7 +555,7 @@ public void initialize() {
 		
 		altura.textProperty().addListener((observable, oldValue, newValue) -> {
 			try{
-				consta[1]=Float.parseFloat(((String)altura.getText()));
+				consta[1]=Double.parseDouble(((String)altura.getText()));
 				consta[3]=consta[1]*consta[2];
 				volumen.setText(String.valueOf(consta[3]));
 				volVm.setText(String.valueOf(consta[3]));
@@ -569,7 +569,7 @@ public void initialize() {
 		
 		pesoAnillo.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
-				consta[4]=Float.parseFloat(((String)pesoAnillo.getText()));
+				consta[4]=Double.parseDouble(((String)pesoAnillo.getText()));
 				wmnatural.setText( String.valueOf(antPrueba[0]-consta[4]));		
 				despPrueba[1]=despPrueba[0]-consta[4];
 				despPrueba[2]=consta[4]+antPrueba[3];
@@ -587,7 +587,7 @@ public void initialize() {
 	///Para calcular la humedad inicial
 		wm.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
-				humIni[0]=Float.parseFloat(((String)wm.getText()));
+				humIni[0]=Double.parseDouble(((String)wm.getText()));
 				humIni[3]=((humIni[0]-humIni[1])/(humIni[1]-humIni[2]))*100;
 				wpercent.setText(String.valueOf(humIni[3]));
 			}catch (Exception e) {
@@ -598,7 +598,7 @@ public void initialize() {
 		
 		ws.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
-				humIni[1]=Float.parseFloat(((String)ws.getText()));
+				humIni[1]=Double.parseDouble(((String)ws.getText()));
 				humIni[3]=((humIni[0]-humIni[1])/(humIni[1]-humIni[2]))*100;
 				wpercent.setText(String.valueOf(humIni[3]));
 			}catch (Exception e) {
@@ -608,7 +608,7 @@ public void initialize() {
 		
 		wf.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
-				humIni[2]=Float.parseFloat(((String)wf.getText()));
+				humIni[2]=Double.parseDouble(((String)wf.getText()));
 				humIni[3]=((humIni[0]-humIni[1])/(humIni[1]-humIni[2]))*100;
 				wpercent.setText(String.valueOf(humIni[3]));
 			}catch (Exception e) {
@@ -620,12 +620,12 @@ public void initialize() {
 		//excepción
 		wpercent.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				antPrueba[2]=Float.parseFloat((String)wpercent.getText());
+				antPrueba[2]=Double.parseDouble((String)wpercent.getText());
 				hum.setText(wpercent.getText());
 				
 				antPrueba[3]=antPrueba[1]/(1+antPrueba[2]/100);
 				wmnaturals.setText(String.valueOf(antPrueba[3]));
-				float ss=Float.parseFloat(((String)Ss.getText()));
+				double ss=Double.parseDouble(((String)Ss.getText()));
 				antPrueba[4]=antPrueba[3]/ss;
 				volmnatural.setText(String.valueOf(antPrueba[4]));
 				antPrueba[5]=antPrueba[3]*antPrueba[2]/100;//Peso(Ww)
@@ -638,7 +638,7 @@ public void initialize() {
 		/////////////////////ANTES DE LA PRUEBA////////////////////////////
 		wma.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
-				antPrueba[0]=Float.parseFloat(((String)wma.getText()));
+				antPrueba[0]=Double.parseDouble(((String)wma.getText()));
 				antPrueba[1]=antPrueba[0]-consta[4];
 				wmnatural.setText( String.valueOf(antPrueba[1]));
 				
@@ -648,7 +648,7 @@ public void initialize() {
 				
 				gm.setText(String.valueOf(antPrueba[1]/consta[3]));//modifica valor en clasificacion 
 				
-				float ss=Float.parseFloat(((String)Ss.getText()));
+				double ss=Double.parseDouble(((String)Ss.getText()));
 				antPrueba[4]=antPrueba[3]/ss;
 				volmnatural.setText(String.valueOf(antPrueba[4]));
 				antPrueba[5]=antPrueba[3]*antPrueba[2]/100;//Peso(Ww)
@@ -672,13 +672,13 @@ public void initialize() {
 		//////////////DESPUES DE LA PRUEBA////////////////////////////////
 		wmsa.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
-				despPrueba[0]=Float.parseFloat(((String)wmsa.getText()));
+				despPrueba[0]=Double.parseDouble(((String)wmsa.getText()));
 				despPrueba[1]=despPrueba[0]-consta[4];
 				despPrueba[2]=consta[4]+antPrueba[3];
 				despPrueba[3]=antPrueba[3];
 				despPrueba[4]=((despPrueba[1]-despPrueba[3])/despPrueba[3])*100;
 				despPrueba[5]=consta[2]*1;//falta vincularlo con resultados
-				despPrueba[6]=despPrueba[3]/Float.parseFloat(((String)Ss.getText()));
+				despPrueba[6]=despPrueba[3]/Double.parseDouble(((String)Ss.getText()));
 				despPrueba[7]=despPrueba[5]-despPrueba[6];
 				despPrueba[8]=1;//datos de resultados
 				despPrueba[10]=despPrueba[0]-despPrueba[2];
@@ -729,8 +729,8 @@ public void initialize() {
 		});
 		i2.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in1.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i2.getText()));
+				double c=Double.parseDouble(((TextField)in1.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i2.getText()));
 				((TextField)in2.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -739,8 +739,8 @@ public void initialize() {
 		
 		i3.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in2.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i3.getText()));
+				double c=Double.parseDouble(((TextField)in2.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i3.getText()));
 				((TextField)in3.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -749,8 +749,8 @@ public void initialize() {
 		
 		i4.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in3.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i4.getText()));
+				double c=Double.parseDouble(((TextField)in3.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i4.getText()));
 				((TextField)in4.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -759,8 +759,8 @@ public void initialize() {
 		
 		i5.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in4.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i5.getText()));
+				double c=Double.parseDouble(((TextField)in4.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i5.getText()));
 				((TextField)in5.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -768,8 +768,8 @@ public void initialize() {
 		});
 		i6.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in5.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i6.getText()));
+				double c=Double.parseDouble(((TextField)in5.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i6.getText()));
 				((TextField)in6.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -778,8 +778,8 @@ public void initialize() {
 		
 		i7.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in6.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i7.getText()));
+				double c=Double.parseDouble(((TextField)in6.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i7.getText()));
 				((TextField)in7.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -788,8 +788,8 @@ public void initialize() {
 		
 		i8.textProperty().addListener((observable,oldValue,newValue)->{
 			try{
-				float c=Float.parseFloat(((TextField)in7.getChildren().get(5)).getText());
-				float sum=c+(Float.parseFloat(i8.getText()));
+				double c=Double.parseDouble(((TextField)in7.getChildren().get(5)).getText());
+				double sum=c+(Double.parseDouble(i8.getText()));
 				((TextField)in8.getChildren().get(5)).setText(String.valueOf(sum));
 			}catch(Exception e){
 				System.out.println(e.getMessage());
@@ -850,9 +850,9 @@ public void initialize() {
 				int tama=(temp.get(1).getChildren().size()-1);
 				t.textProperty().addListener((observable,lastValue,newValue)->{
 					try{
-						float ini=Float.parseFloat(((TextField)temp.get(1).getChildren().get(7)).getText());
+						double ini=Double.parseDouble(((TextField)temp.get(1).getChildren().get(7)).getText());
 						
-						float rest=ini-(Float.parseFloat(t.getText()));
+						double rest=ini-(Double.parseDouble(t.getText()));
 						
 						((TextField)temp.get(1).getChildren().get(tama-2)).setText(String.valueOf(rest));
 						((TextField)temp.get(1).getChildren().get(tama-2)).setEditable(false);
@@ -883,8 +883,8 @@ public void initialize() {
 					//int cantidad=((tam-16)/7)+1;
 					//System.out.println(cantidad);
 					for (int i=14; i<temp.get(1).getChildren().size();i+=6){
-						float m=Float.parseFloat(((TextField)temp.get(1).getChildren().get(i)).getText());
-						float df=Float.parseFloat(((TextField)temp.get(1).getChildren().get(i+3)).getText());
+						double m=Double.parseDouble(((TextField)temp.get(1).getChildren().get(i)).getText());
+						double df=Double.parseDouble(((TextField)temp.get(1).getChildren().get(i+3)).getText());
 						s.getData().add(new XYChart.Data(m,df));
 					}
 				}catch(Exception ec){
@@ -899,8 +899,8 @@ public void initialize() {
 					//int cantidad=((tam-16)/7)+1;
 					//System.out.println(cantidad);
 					for (int i=14; i<temp.get(1).getChildren().size();i+=6){
-						float m=Float.parseFloat(((TextField)temp.get(1).getChildren().get(i)).getText());
-						float df=Float.parseFloat(((TextField)temp.get(1).getChildren().get(i+3)).getText());
+						double m=Double.parseDouble(((TextField)temp.get(1).getChildren().get(i)).getText());
+						double df=Double.parseDouble(((TextField)temp.get(1).getChildren().get(i+3)).getText());
 						s.getData().add(new XYChart.Data(m,df));
 					}
 				}catch(Exception ec){
@@ -943,7 +943,7 @@ public void initialize() {
 		c.textProperty().addListener((observable, oldValue, newValue)->{
 			try{
 				((TextField)d2.getChildren().get(7)).setText(c.getText());
-				float res = - Float.parseFloat( ((TextField)d.getChildren().get(7)).getText() ) + Float.parseFloat(c.getText());
+				double res = - Double.parseDouble( ((TextField)d.getChildren().get(7)).getText() ) + Double.parseDouble(c.getText());
 				((TextField)d.getChildren().get(16)).setText(String.valueOf(res));
 				((TextField)d.getChildren().get(17)).setText(String.valueOf(res));
 			}catch(Exception e){
